Fix item image loading
